Move compiledir and compilelock to compile module
And specify their purpose and public API.
