+++ doc/extending/pipeline.txt
@@ -116,7 +116,7 @@ case if ``borrow`` was True, the thunk would be allowed to reuse (or
     crashes with paralell execution of some scripts). This mechanism is
     currently enabled by default, but if it causes any problem it may be
     disabled using the function
-    ``theano.gof.compilelock.set_lock_status(..)``.
+    ``theano.compile.compilelock.set_lock_status(..)``.
 
 Step 4 - Wrap the thunk in a pretty package
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
