+++ theano/link/c/cutils.py
@@ -2,8 +2,8 @@ import errno
 import os
 import sys
 
+from theano.compile.compilelock import get_lock, release_lock
 from theano.configdefaults import config
-from theano.gof.compilelock import get_lock, release_lock
 from theano.link.c import cmodule
