+++ theano/scan/scan_perform_ext.py
@@ -17,8 +17,8 @@ from importlib import reload
 import numpy as np
 
 import theano
+from theano.compile.compilelock import get_lock, release_lock
 from theano.configdefaults import config
-from theano.gof.compilelock import get_lock, release_lock
 from theano.link.c import cmodule
